Update dropdown inputs based on radio button selection

Updates searchable dropdown inputs for species selection based on the user's choice of searching by vernacular name or scientific name. When a user selects a species in one dropdown, this function updates the corresponding value in the other dropdown.

Usage

updateDropdownBasedOnRadioBtnValues(
  radioBtn_search_byName,
  radioBtn_search_byNameID_ver,
  id_ver_name,
  scientificName_str,
  vernacularName_str,
  id_selected_sci_name,
  radioBtn_search_byNameID_sci,
  id_sci_name,
  id_selected_ver_name,
  session
)

Arguments

radioBtn_search_byName

The current value of the radio button that determines the search method (vernacular or scientific name)

radioBtn_search_byNameID_ver

The value of the radio button that corresponds to searching by vernacular name

id_ver_name

The input value from the vernacular name dropdown

scientificName_str

Constant string identifier for scientific name, typically "scientificName"

vernacularName_str

Constant string identifier for vernacular name, typically "vernacularName"

id_selected_sci_name

The input value from the matching scientific name dropdown

radioBtn_search_byNameID_sci

The value of the radio button that corresponds to searching by scientific name

id_sci_name

The input value from the scientific name dropdown

id_selected_ver_name

The input value from the matching vernacular name dropdown

session

The Shiny session object