Update Dropdown Based on Radio Button

updateDropdownBasedOnRadioBtnValues will updates input values of searchable dropdowns selecting species byVernacular name and by scientific name based on radio button for Search species by Vernacuar Name or Scientific Name

Usage

updateDropdownBasedOnRadioBtnValues(
  radioBtn_search_byName,
  radioBtn_search_byNameID_ver,
  id_ver_name,
  scientificName_str,
  vernacularName_str,
  id_selected_sci_name,
  radioBtn_search_byNameID_sci,
  id_sci_name,
  id_selected_ver_name,
  session
)

Arguments

radioBtn_search_byName

and radioBtn_search_byNameID_ver,radioBtn_search_byNameID_sci are Search by Name radio button input value and specific choice like Vernacular Name or Scientific Name. id_ver_name, id_sci_name are input values of Search Species by Vernacular Name or Scientific Name scientificName_str, vernacularName_str constant string "scientificName" and "vernacularName" id_selected_sci_name, id_selected_ver_name,session are input values of Matching Scientific Name and Matching Vernacular Name and session of shiny app.